ajout contribution
TO DO : ajouter collecte

+++ AssoFlex/Models/Contribution.cs
@@ -6,7 +6,7 @@ namespace AssoFlex.Models
     {
         public int Id{ get; set; }
              
-        public int MontantContribution { get; set; }
+        public double MontantContribution { get; set; }
         public DateTime DateContribution { get; set; }
         public int utilisateurId { get; set; }
         public Utilisateur utilisateur { get; set; }
